+++ Models/Study/Student.cs
@@ -2,11 +2,11 @@ namespace Dotnet.Models
 {
 	public class Student
 	{
-		public int Id { get; set; }
-		public int UserId { get; set;}	
+		public ulong Id { get; set; }
+		public ulong UserId { get; set;}	
 		public bool IsLearns { get; set; }
 		public bool IsOnline { get; set; }
-		public int StudySubgroupId { get; set; }
+		public ulong StudySubgroupId { get; set; }
 		public StudySubgroup StudySubgroup { get; set; }
 	}
 }
