+++ Models/Study/StudySubgroup.cs
@@ -5,10 +5,10 @@ namespace Dotnet.Models
 {
 	public class StudySubgroup
 	{
-		public int Id { get; set; }
+		public ulong Id { get; set; }
 		public string Name { get; set; }
 		public string Code { get; set; }
-		public int StudyGroupId { get; set; }
+		public ulong StudyGroupId { get; set; }
 		public StudyGroup StudyGroup { get; set; }
 		public List<StudySubgroupWork> StudySubgroupWork { get; set; } = new List<StudySubgroupWork>();
 		public List<Student> Students { get; set; } = new List<Student>();
