+++ Models/Study/StudySubgroupWork.cs
@@ -2,10 +2,10 @@ namespace Dotnet.Models.Study
 {
 	public class StudySubgroupWork 
 	{
-		public int Id { get; set; }
-		public int WorkId { get; set; }
+		public ulong Id { get; set; }
+		public ulong WorkId { get; set; }
 		public Work Work { get; set; }
-		public int StudySubgroupId { get; set; }
+		public ulong StudySubgroupId { get; set; }
 		public StudySubgroup StudySubgroup { get; set; }
 	}
 }
