+++ ViewModels/WebApp/Student/EditStudentViewModel.cs
@@ -5,16 +5,16 @@ namespace Dotnet.ViewModels.WebApp.Student
 	public class EditStudentViewModel
 	{
 		[Required]
-		public int Id { get; set; }
+		public ulong Id { get; set; }
 
 		[Required(ErrorMessage = "Укажите пользователя.")]
-		public int UserId { get; set; }	
+		public ulong UserId { get; set; }	
 		
 		[Required(ErrorMessage = "Укажите статус.")]
 		public bool IsLearns { get; set; }
 
 		[Required(ErrorMessage = "Укажите учебную подгруппу.")]
-		public int StudySubgroupId { get; set; }
+		public ulong StudySubgroupId { get; set; }
 
 		public bool IsOnline { get; set; } = false;
 	}
