+++ ViewModels/WebApp/StudyGroup/EditStudyGroupViewModel.cs
@@ -5,7 +5,7 @@ namespace Dotnet.ViewModels.WebApp.StudyGroup
     public class EditStudyGroupViewModel
     {
 		[Required]
-		public int Id { get; set; }
+		public ulong Id { get; set; }
 
 		[Required(ErrorMessage = "Укажите название учебной группы")]
 		public string Name { get; set; }
@@ -19,9 +19,9 @@ namespace Dotnet.ViewModels.WebApp.StudyGroup
 		public System.DateTime DateEnd { get; set; }
 
 		[Required(ErrorMessage = "Укажите специальность")]
-		public int FormEducationId { get; set; }
+		public ulong FormEducationId { get; set; }
 
 		[Required(ErrorMessage = "Укажите форму обучения")]
-		public int SpecialtyId { get; set; }
+		public ulong SpecialtyId { get; set; }
     }
 }
