+++ ViewModels/WebApp/StudyGroup/EditStudySubgroupViewModel.cs
@@ -5,10 +5,10 @@ namespace Dotnet.ViewModels.WebApp.StudyGroup
     public class EditStudySubgroupViewModel
     {
 		[Required]
-		public int Id { get; set; }
+		public ulong Id { get; set; }
 
 		[Required(ErrorMessage = "Укажите ID учебной группы")]
-		public int StudyGroupId { get; set; }
+		public ulong StudyGroupId { get; set; }
 
 		[Required(ErrorMessage = "Укажите название учебной подгруппы")]
 		public string Name { get; set; }
