+++ ViewModels/WebApp/Teacher/EditTeacherViewModel.cs
@@ -5,10 +5,10 @@ namespace Dotnet.ViewModels.WebApp.Teacher
 	public class EditTeacherViewModel
 	{
        	[Required]
-        public int Id { get; set; }
+        public ulong Id { get; set; }
 
         [Required(ErrorMessage = "Введите идентификатор пользователя")]
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
 
         public string Post { get; set; }
