+++ ViewModels/WebApp/Teacher/TeacherViewModel.cs
@@ -5,7 +5,7 @@ namespace Dotnet.ViewModels.WebApp.Teacher
     public class TeacherViewModel
     {
         [Required(ErrorMessage = "Укажите ID пользователя")]
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
          
         public string Post { get; set; }
